third_party/WebKit/LayoutTests/editing/inserting/typing-at-end-of-line.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
const style='font: 12px lucida\\ grande; width: 871px; height: 100px;';
test(() => assert_selection(
  [
    `<div contenteditable style="${style}">`,
      'and Anna, you can do your regular weekly report run, and I\'ll extract ',
      'the key sections into this. Andrew and Mark, I\'d like to ',
      '|include any new bits you',
    '</div>',
  ].join(''),
  selection => selection.document.execCommand('insertText', false, ' have'),
  [
    `<div contenteditable style="${style}">`,
      'and Anna, you can do your regular weekly report run, and I\'ll extract ',
      'the key sections into this. Andrew and Mark, I\'d like to\u00A0',
      ' have|include any new bits you',
    '</div>',
  ].join('')),
  'Text does not split on white space when typing to the edge of window');
</script>
